fix: sizing and suspense
