+++ clients/search-component/src/TrieveModal/Search/SearchMode.tsx
@@ -1,4 +1,4 @@
-import React, { Suspense, useEffect, useMemo } from "react";
+import React, { useEffect, useMemo } from "react";
 import { DocsItem } from "./DocsItem";
 import { ModalProps, useModalState } from "../../utils/hooks/modal-context";
 import {
@@ -10,7 +10,6 @@ import {
 import { ProductItem } from "./ProductItem";
 import { ProductGroupItem } from "./ProductGroupItem";
 import { PdfItem } from "./PdfItem";
-import { ModeSwitch } from "../ModeSwitch";
 import { cn } from "../../utils/styles";
 import { SearchInput } from "./SearchInput";
 import { GoToChatPrompt } from "./GoToChatPrompt";
@@ -21,7 +20,6 @@ export const SearchMode = () => {
     results,
     loadingResults,
     query,
-    setOpen,
     requestID,
     inputRef,
     open,
@@ -117,18 +115,7 @@ export const SearchMode = () => {
   const hasQuery = imageUrl || query || audioBase64;
 
   return (
-    <Suspense fallback={<div className="tv-hidden"></div>}>
-      {!props.inline && (
-        <div className="mode-switch-wrapper tv-flex tv-items-center tv-px-2 tv-gap-2 tv-justify-end tv-mt-2 tv-font-medium ${mode}">
-          <ModeSwitch />
-          <div
-            className={`tv-text-xs tv-rounded-md !tv-bg-transparent tv-flex !hover:bg-tv-zinc-200 tv-px-2 tv-justify-end tv-items-center tv-p-2 tv-gap-0.5 tv-cursor-pointer ${props.type}`}
-            onClick={() => setOpen(false)}
-          >
-            <CloseIcon />
-          </div>
-        </div>
-      )}
+    <>
       <SearchInput />
       {resultsLength && props.chat && imageUrl.length == 0 ? (
         <GoToChatPrompt />
@@ -155,7 +142,7 @@ export const SearchMode = () => {
           <p className="">Searching...</p>
         </div>
       )}
-    </Suspense>
+    </>
   );
 };
 
@@ -178,30 +165,6 @@ const NoResults = ({ props, query }: { props: ModalProps; query: string }) => {
         </p>
       )}
     </div>
-  );
-};
-
-const CloseIcon = () => {
-  return (
-    <>
-      <svg
-        className="close-icon"
-        xmlns="http://www.w3.org/2000/svg"
-        width="24"
-        height="24"
-        viewBox="0 0 24 24"
-        fill="none"
-        stroke="currentColor"
-        strokeWidth="2"
-        strokeLinecap="round"
-        strokeLinejoin="round"
-      >
-        <path stroke="none" d="M0 0h24v24H0z" fill="none" />
-        <line x1="18" y1="6" x2="6" y2="18" />
-        <line x1="6" y1="6" x2="18" y2="18" />
-      </svg>
-      <span>Close</span>
-    </>
   );
 };
