+++ clients/search-component/src/utils/hooks/useSizing.ts
@@ -6,13 +6,14 @@ const popupClass = cva(
     "tv-max-h-[40vh]",
     "tv-min-w-[90vw]",
     "sm:tv-max-w-[800px]",
-    "tv-top-[calc(40%-30vh)] tv-left-[50%] tv-shadow-2xl tv-fixed -tv-translate-x-[50%]",
+    "tv-top-[calc(40%-30vh)]",
+    "tv-left-[50%] tv-shadow-2xl tv-fixed -tv-translate-x-[50%]",
     "tv-w-[90vw]",
   ],
   {
     variants: {
       type: {
-        ecommerce: ["tv-px-4", "tv-top-[0px]"],
+        ecommerce: ["tv-px-4", "!tv-top-[0px]"],
         docs: [],
         pdf: [],
       },
@@ -23,15 +24,14 @@ const popupClass = cva(
 const inlineClass = cva(
   [
     "tv-max-h-[40vh]",
-    "tv-w-[90vw]",
     "sm:tv-max-w-[800px]",
-    "!tv-min-w-full tv-max-w-sm !tv-w-full",
+    "!tv-min-w-full tv-max-w-sm tv-w-full",
   ],
   {
     variants: {
       type: {
         ecommerce: [
-          "tv-top-1 tv-w-[95vw] tv-rounded-lg",
+          "tv-top-1 tv-w-[95vw] tv-min-w-full tv-rounded-lg",
           "tv-px-0 tv-pt-0 tv-max-w-full",
         ],
         docs: [],
